@@ -3,10 +3,6 @@
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
-  "engines": {
-    "node": "8.11.1",
-    "npm": "6.1.0"
-  },
   "scripts": {
     "build": "nuxt build",
     "start": "nuxt start",
@@ -23,30 +19,29 @@
   },
   "homepage": "https://github.com/imere/nightlife-coordination#readme",
   "dependencies": {
-    "axios": "^0.18.0",
-    "connect-mongo": "^2.0.1",
-    "es6-promise": "^4.2.4",
-    "express": "^4.16.3",
-    "express-session": "^1.15.6",
-    "mongoose": "^5.1.4",
-    "nuxt": "^1.4.1",
-    "vue-notification": "^1.3.10",
+    "@nuxt/webpack": "^2.14.12",
+    "axios": "^0.21.1",
+    "connect-mongo": "^3.2.0",
+    "express": "^4.17.1",
+    "express-session": "^1.17.1",
+    "mongoose": "^5.11.14",
+    "nuxt": "^2.14.12",
+    "vue-notification": "^1.3.20",
     "yelp-fusion": "^2.0.3"
   },
   "devDependencies": {
-    "ajv": "^6.5.1",
-    "ajv-keywords": "^3.2.0",
-    "babel-eslint": "^8.2.3",
-    "eslint": "^4.19.1",
-    "eslint-config-standard": "^11.0.0",
-    "eslint-loader": "^2.0.0",
-    "eslint-plugin-html": "^4.0.3",
-    "eslint-plugin-import": "^2.12.0",
-    "eslint-plugin-node": "^6.0.1",
-    "eslint-plugin-promise": "^3.8.0",
-    "eslint-plugin-standard": "^3.1.0",
-    "node-sass": "^4.9.0",
-    "sass-loader": "^7.0.3",
-    "webpack": "^3.12.0"
+    "@nuxtjs/eslint-module": "^3.0.2",
+    "@nuxtjs/style-resources": "^1.0.0",
+    "babel-eslint": "^10.1.0",
+    "eslint": "^7.18.0",
+    "eslint-config-standard": "^16.0.2",
+    "eslint-plugin-html": "^6.1.1",
+    "eslint-plugin-import": "^2.22.1",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-promise": "^4.2.1",
+    "eslint-plugin-standard": "^5.0.0",
+    "eslint-plugin-vue": "^7.5.0",
+    "node-sass": "^5.0.0",
+    "sass-loader": "^10.1.1"
   }
 }
